update18

+++ src/app/starter/starter-content/starter-content.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit,HostListener, ANALYZE_FOR_ENTRY_COMPONENTS  } from '@angular/core';
+import { Component, OnInit,HostListener,ANALYZE_FOR_ENTRY_COMPONENTS  } from '@angular/core';
 import { MessageService } from '../../services/message.service';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators, NgForm } from '@angular/forms';
+import { analyzeAndValidateNgModules } from '@angular/compiler';
+import { last } from '@angular/router/src/utils/collection';
 
 declare var $: any;
 
@@ -23,29 +25,70 @@ export class StarterContentComponent implements OnInit {
   msgInputId:any;
   inputString = "ipt_";
   updateMsg : any;
-  display='none';
-   
-  bkData:string;
+  
+  displayMsgFlag:boolean = false;
+  newMsgVal:string='';
+  newMsgFlag:string='';
+  newMsgID:any;
 
+  display='none';	
+  bkData:string;
+ 
   ngOnInit() { 
       this.getAllMessage(); 
   }
- 
+
+  addMessage(msgfalg: NgForm){
+    //setting new message to active
+    if(this.displayMsgFlag == true){
+      this.newMsgFlag = "Y";
+    }else{
+      this.newMsgFlag = "N";
+    }
+    
+    //setting new message ID
+    this.newMsgID = Math.max.apply(Math,this.messageList.map(function(o) { return o.msgID; }));
+    //console.log(this.newMsgID)
+    this.newMsgID = this.newMsgID + 1
+    //console.log(this.newMsgID)
+    this.newMsgVal = msgfalg.value.inewmsg;
+
+    //console.log(this.newMsgVal)
+    //console.log(this.newMsgID)
+    //console.log(this.newMsgFlag)
+
+    //setting new message data
+    const msgBody = {msgID: this.newMsgID, msg: this.newMsgVal, displayMsg: this.newMsgFlag};
+  
+    //console.log(msgfalg);
+    //console.log(msgBody);
+    //sending to message.service for post
+    this.messageService.addMessage(msgBody).subscribe((result) => {        
+      console.log(result);
+      this.getAllMessage();
+    }, err  => this.errorMsg = <any>err);   
+    console.log(this.messageList)
+    msgfalg.reset();
+  }
+
+  getOrignalVal(data:any,id,i){ 
+     this.originalVal = data;
+     this.msgInputId = id;
+     this.updateMsg = i;     
+  }
   @HostListener('document:keydown', ['$event']) onKeydownHandler(event: KeyboardEvent) {
     if (event.key === "Escape") {  
       this.messageList[this.updateMsg].msg = this.originalVal;
     }
   }
 
-  lenghtValidate(data:any){ 
-    console.log(data);
-    if(data.length >=75){
-      alert("Message Should Not Be More Than 75 Characters.");
-      return false;
-    }
-    
-    return true;
-  }  
+  // lenghtValidate(data:any){
+  //   if(data.length >=70){
+  //     alert("Message Should Not Be More Than 70 Characters.");
+  //     return false;
+  //   }
+  //   return true;
+  // } 
 
   getAllMessage(){
     this.messageService.getAllMessage().subscribe((messageList) => { 
@@ -54,18 +97,14 @@ export class StarterContentComponent implements OnInit {
       console.log(this.messageList);
     }, err  => this.errorMsg = <any>err);
   }
-
   
   updateMessage(data:any,i){ 
     this.errorMsg = '';
     let bkDataObj = JSON.parse(this.bkData); 
-    let updatedStr = data.msg.replace(/ /g, "");
-    console.log(updatedStr);
-
+    let updatedStr = data.msg.replace(/ /g, ""); 
     $("#info_"+data.msgID).hide();
     $("#save_"+data.msgID).hide();
-    $("#error_"+data.msgID).hide();
-
+    $("#error_"+data.msgID).hide(); 
     if(bkDataObj[i].msg == updatedStr){   
      $("#info_"+data.msgID).show();
     }else{        
@@ -77,44 +116,25 @@ export class StarterContentComponent implements OnInit {
         $("#error_"+data.msgID).hide();
 
       }, err  => this.errorMsg = <any>err); 
-
-      console.log(this.errorMsg);
-
+ 
       if(this.errorMsg){
         $("#info_"+data.msgID).hide();
         $("#save_"+data.msgID).hide();
         $("#error_"+data.msgID).show();
       }
  
-    }
-
-    
-     
+    } 
 
   }
-
-  persistMsgTmp(data:any){
-    this.display='block'; 
+ 
+  deleteMessage(data:any){
     console.log(data);
-    window.localStorage.setItem("dataForDelete",JSON.stringify(data));
-  }
-
-  removeDeleteData(){
-    window.localStorage.removeItem("dataForDelete");
-    this.display='none'; 
-  }
-
-  deleteMessage(){ 
-
-    let data = JSON.parse(window.localStorage.getItem("dataForDelete"));
-    this.messageService.deleteMessage(data).subscribe((result) => {        
-      console.log(result);
-      this.getAllMessage();
-      this.removeDeleteData(); 
-    }, err  => this.errorMsg = <any>err);
-
-    this.display='none'; 
-    
+    if(window.confirm("Are you sure?")) {
+      this.messageService.deleteMessage(data).subscribe((result) => {        
+        console.log(result);
+        this.getAllMessage();
+      }, err  => this.errorMsg = <any>err);
+    }
   }
 
   selectedSection(item:any){
@@ -128,7 +148,6 @@ export class StarterContentComponent implements OnInit {
     } 
   }
 
-
   getStatus(data:any){ 
 
       if(data.displayMsg == "N"){ 
@@ -136,7 +155,6 @@ export class StarterContentComponent implements OnInit {
       } else{
         return true;
       }
-      
   }
 
   getStatusChk(ev,data){ 
@@ -144,8 +162,8 @@ export class StarterContentComponent implements OnInit {
       data.displayMsg = "Y";
     }else{
       data.displayMsg = "N";
-    }
-    
+    }    
   }
-
+  
+  
 }
